Move trustregion subsolvers to separate folder.
