+++ src/estimagic/optimization/trustregion/_conjugate_gradient_quadratic.py
