+++ src/estimagic/optimization/trustregion/_steihaug_toint_quadratic.py
