+++ src/estimagic/optimization/trustregion/_trsbox_quadratic.py
