+++ src/estimagic/optimization/trustregion/gqtpar_quadratic.py
