+++ src/estimagic/optimization/trustregion/linear_subsolvers.py
